complete application

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
             'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!') ;
@@ -50,6 +52,11 @@ class ListingController extends Controller
 
     //Show Edit Form
     public function edit (Listing $listing){
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
@@ -77,8 +84,18 @@ class ListingController extends Controller
 
         //Delete Listing
         public function destroy(Listing $listing){
+             //Make sure logged in user is owner
+             if($listing->user_id != auth()->id()){
+                abort(403, 'Unauthorized Action');
+            }
+
             $listing->delete();
             return redirect('/')->with('message', 'Listing deleted successfully');
+        }
+
+        //Manage Listing
+        public function manage(){
+            return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
         }
 }
